Switch HashTable::find() to pointer-like interface.
Before this commit, find() returned an iterator. This posed a problem
for compound hash tables that do not support iteration, making them
implement a hack-ish iterator-like wrapper with fake operator ==() that
only compared to end().

Instead of that, return a pointer-like object to a (key, mapped &) pair. In
the simplest case of normal HashTable, return a raw pointer to the Cell
object. In hash tables that do not have ready (key, mapped &) pairs,
such as FixedHashTable, return a std::optional over a wrapper object.
Use the same wrapper we use in the corresponding iterator class.